andfun-kotlin-sleep-tracker/app/src/androidTest/java/com/example/android/trackmysleepquality/SleepDatabaseTest.kt
/*
 * Copyright 2018, The Android Open Source Project
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package com.example.android.trackmysleepquality

import androidx.room.Room
import androidx.test.ext.junit.runners.AndroidJUnit4
import androidx.test.platform.app.InstrumentationRegistry
import com.example.android.trackmysleepquality.database.SleepDatabase
import com.example.android.trackmysleepquality.database.SleepDatabaseDao
import com.example.android.trackmysleepquality.database.SleepNight
import org.junit.Assert.assertEquals
import org.junit.After
import org.junit.Before
import org.junit.Test
import org.junit.runner.RunWith
import java.io.IOException

/**
 * This is not meant to be a full set of tests. For simplicity, most of your samples do not
 * include tests. However, when building the Room, it is helpful to make sure it works before
 * adding the UI.
 */

// @RunWith sets up the test runner we are using
@RunWith(AndroidJUnit4::class)
class SleepDatabaseTest {

    private lateinit var sleepDao: SleepDatabaseDao
    private lateinit var db: SleepDatabase

    // @Before annotation makes method a setup method, so it runs before the tests
    @Before
    fun createDb() {
        val context = InstrumentationRegistry.getInstrumentation().targetContext
        // Using an in-memory database because the information stored here disappears when the
        // process is killed.
        db = Room.inMemoryDatabaseBuilder(context, SleepDatabase::class.java)
                // Allowing main thread queries, just for testing.
                .allowMainThreadQueries()
                .build()
        sleepDao = db.sleepDatabaseDao
    }

    // Function annotated with @after executes after testing is finished
    // This is where we close the database connection
    @After
    @Throws(IOException::class)
    fun closeDb() {
        db.close()
    }

    // This is a test method annotated with @Test
    @Test
    @Throws(Exception::class)   // Throw an exception if test fails
    fun insertAndGetNight() {
        val night = SleepNight()    // Create a new SleepNight instance
        sleepDao.insert(night)      // Try inserting in db
        val tonight = sleepDao.getTonight() // Get latest sleep night data
        assertEquals(tonight?.sleepQuality, -1)
        // Check sleep quality is -1,
        // since we did not specify a sleep quality in SleepNight()
    }
}

